refactor(i18n): split the message vocabulary into per-domain modules
`message.rs` was one flat `UserMessage` enum of 200 variants kept in three
hand-maintained parallel lists — the enum, the `localized()` match, and the
sample list — in a single 1,889-line file that every feature has to touch.
PR #122 had to invent a `Status*` prefix to namespace its 55 additions, which
is the module boundary asking to exist; ~385 raw print sites still waiting to
be converted would have roughly doubled the file.

Split it into nine per-domain types unified behind one `Message` trait
carrying the two renderings. The sinks (`Terminal`, `fail`), the gettext
plumbing (`init`, `translate`, `fill`) and the trait stay shared in
`message/mod.rs`; each domain owns its variants, its `localized()` arms and
its own sample walk:

    access    10   config load, privilege, which backend answered
    face      32   enroll / test / remove / clear
    status    55   the `facelock status` report's prose
    notify     4   notification bodies (rendered NotifyEvent)
    setup     47   wizard spine, step failures, closing summary
    device    12   camera and inference-device selection
    download   9   model download and verification
    system    11   daemon unit and `facelock` group setup
    pam       20   /etc/pam.d/* configuration

The domains follow the section banners the file had already grown, so the
boundaries are the ones the vocabulary picked for itself rather than a
taxonomy imposed on it. The largest file drops from 1,889 lines to 336.

Variant names are unchanged. Both renderings are byte-identical for all 200
messages: `localized()` because container tests grep it and the status
renderer has a golden byte-exact test, and `machine()` because it is derived
from `Debug` and feeds the tracing event stream. `just pot` now globs the
domain modules (xgettext takes multiple inputs) and extracts the same 200
msgids; only the `#:` source references move.

The sample list that feeds the placeholder sweep is now a wildcard-free
`match` per domain, so a variant added without a sample does not compile —
previously the list could silently fall behind the enum.
